Feat: Make the frontend recognise left mouse clicks
I used event.type == pygame.MOUSEBUTTON to check if the mouse has been
clicked. If a mouse button has been clicked then the event will have an
attribute called button. The value of this (whether 1 or 2) will tell
you if the left or the right mouse button has been clicked. There is
also an attribute called pos which gives the position of the click but
only if it is within the escrren area.