Feat : Add map generation to backend, button clicking to frontend.
The map is implemented as a 2d list of dictionaries. Each box that the
player can click on in the minefield corresponds to exactly one
dictionary in theMap data structure in the backend. Each of these
dictionaries describe the current state of the box. The map is redrawn
after each event based on the values of all these boxes.

Currently the frontend is just one black screen. Added logical sections
in the screen that correspond to buttons